ANDROID_CV/AiModel/requirements.txt
ultralytics>=8.0.0
torch
torchvision
Pillow
# TensorFlow is required for TFLite export
tensorflow-cpu>=2.14.0
